+++ android/src/main/java/ai/openspace/backgroundupload/NotificationReceiver.kt
@@ -1,4 +1,4 @@
-package com.vydia.RNUploader
+package ai.openspace.backgroundupload
 
 import android.content.BroadcastReceiver
 import android.content.Context